Hey, so I found an astroquery that sort of works with ESO data!
If you have a user portal (and password) with ESO, there are
queries that allow you to search various ESO missions and return
any science objects within a certain RA/DEC box region. For now
I'm just testing out with gamma Vel. Turns out we get (as expected)
a bunch of stuff from giraffe/uves -- which is what we'd expect!