I finally managed to query the ESO missions. Tried for a few hours to figure out if I could loop them within the ESO class I made, but each time the files just kept overwriting. Never mind, in the end I set up something similar to the VCatalog class where instead of the "catalog", you can type in the name of the "mission". So far I included "giraffe", "uves" and "xshooter" and it's fairly easy to add more.
One thing I couldn't figure out is how to remove the row limit of
50 entries in the astroquery/Vizier tools. I know there's an
explicit way of setting

Vizier.ROW_LIMITS = -1

...but I still keep getting <=50 objects printed to file. We'll have
to take a look at this on Monday I guess.

Also, I'm still running into a few problems with the wcs coordinates
for some of the drz files. I remember we figured out a workaround for
this last time (when I was initially using the .flc files).

Changed the symbol notation to something nicer too.